Decouple command name from repo name
Move the vCenter agent source into cmd/brkt-vcenter-agent.  This removes
the coupling between the name of the repo (top-level directory) and the
name of the executable.

The next step is to rename the repo from "vcenter-agent" to vmware.
This will allow users to run

go get github.com/brkt/vmware

to download the source into their GOPATH and install the brkt-vcenter-
agent binary into $GOPATH/bin.  We'll also have the option of adding
other VMware-related commands to this repo.