Log customer ID
Log the customer ID when starting up.  The vCenter agent pushes data
into a single account.  This allows the user to confirm that VM
properties are being pushed to the correct account.  Also, log the name
and email address when verbose logging is enabled.

Refactor the code that generates an HTTP client and request, so that it
can be used from both callsites.