Handle Content-Type for requests and responses
Set the Content-Type header when sending requests to the Bracket
service.  When handling the response from the customer/self endpoint,
make sure that the content type is application/json.  If the user
accidentally points the agent to a web server (e.g. mgmt.brkt.com
instead of api.mgmt.brkt.com), the server may return a successful HTML
response, which would cause JSON parsing to fail.